Corrected observation search param, converted sample to a list, and created new RequestUtils class, using it where appropriate to avoid repeated code.
