+++ patient-generated-data/src/test/java/gov/va/api/health/patientgenerateddata/RequestUtils.java
@@ -1,0 +1,20 @@
+package gov.va.api.health.patientgenerateddata;
+
+import org.springframework.mock.web.MockHttpServletRequest;
+import org.springframework.web.util.UriComponentsBuilder;
+
+public class RequestUtils {
+
+  public static MockHttpServletRequest requestFromUri(String uri) {
+    var u = UriComponentsBuilder.fromUriString(uri).build();
+    MockHttpServletRequest request = new MockHttpServletRequest();
+    request.setRequestURI(u.getPath());
+    request.setRemoteHost(u.getHost());
+    request.setProtocol(u.getScheme());
+    request.setServerPort(u.getPort());
+    u.getQueryParams()
+        .entrySet()
+        .forEach(e -> request.addParameter(e.getKey(), e.getValue().toArray(new String[0])));
+    return request;
+  }
+}
