+++ patient-generated-data/src/test/java/gov/va/api/health/patientgenerateddata/observation/ObservationControllerTest.java
@@ -1,5 +1,6 @@
 package gov.va.api.health.patientgenerateddata.observation;
 
+import static gov.va.api.health.patientgenerateddata.RequestUtils.requestFromUri;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
@@ -26,9 +27,7 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.ResponseEntity;
-import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.validation.DataBinder;
-import org.springframework.web.util.UriComponentsBuilder;
 
 public class ObservationControllerTest {
   private static final ObjectMapper MAPPER = JacksonMapperConfig.createMapper();
@@ -40,19 +39,6 @@ public class ObservationControllerTest {
           .baseUrl("http://foo.com")
           .r4BasePath("r4")
           .build();
-
-  private static MockHttpServletRequest requestFromUri(String uri) {
-    var u = UriComponentsBuilder.fromUriString(uri).build();
-    MockHttpServletRequest request = new MockHttpServletRequest();
-    request.setRequestURI(u.getPath());
-    request.setRemoteHost(u.getHost());
-    request.setProtocol(u.getScheme());
-    request.setServerPort(u.getPort());
-    u.getQueryParams()
-        .entrySet()
-        .forEach(e -> request.addParameter(e.getKey(), e.getValue().toArray(new String[0])));
-    return request;
-  }
 
   private ObservationController controller() {
     ObservationRepository repo = mock(ObservationRepository.class);
@@ -146,7 +132,7 @@ public class ObservationControllerTest {
 
   @SneakyThrows
   @ParameterizedTest
-  @ValueSource(strings = "?id=1")
+  @ValueSource(strings = "?_id=1")
   void validSearch(String query) {
     ObservationRepository repo = mock(ObservationRepository.class);
     ObservationController controller = controller(repo);
