+++ patient-generated-data/src/test/java/gov/va/api/health/patientgenerateddata/questionnaire/QuestionnaireControllerTest.java
@@ -1,5 +1,6 @@
 package gov.va.api.health.patientgenerateddata.questionnaire;
 
+import static gov.va.api.health.patientgenerateddata.RequestUtils.requestFromUri;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -28,25 +29,10 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.ResponseEntity;
-import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.validation.DataBinder;
-import org.springframework.web.util.UriComponentsBuilder;
 
 public class QuestionnaireControllerTest {
   private static final ObjectMapper MAPPER = JacksonMapperConfig.createMapper();
-
-  private static MockHttpServletRequest _requestFromUri(String uri) {
-    var u = UriComponentsBuilder.fromUriString(uri).build();
-    MockHttpServletRequest request = new MockHttpServletRequest();
-    request.setRequestURI(u.getPath());
-    request.setRemoteHost(u.getHost());
-    request.setProtocol(u.getScheme());
-    request.setServerPort(u.getPort());
-    u.getQueryParams()
-        .entrySet()
-        .forEach(e -> request.addParameter(e.getKey(), e.getValue().toArray(new String[0])));
-    return request;
-  }
 
   private Questionnaire _questionnaire() {
     return Questionnaire.builder().title("x").status(PublicationStatus.active).build();
@@ -116,7 +102,7 @@ public class QuestionnaireControllerTest {
             .build();
     QuestionnaireController controller =
         new QuestionnaireController(pageLinks, mock(QuestionnaireRepository.class));
-    var req = _requestFromUri("http://fonzy.com/r4/Questionnaire" + query);
+    var req = requestFromUri("http://fonzy.com/r4/Questionnaire" + query);
     assertThatExceptionOfType(InvalidRequest.class).isThrownBy(() -> controller.search(req));
   }
 
@@ -140,7 +126,7 @@ public class QuestionnaireControllerTest {
                     List.of(QuestionnaireEntity.builder().build().id("1").payload("{ \"id\": 1}")),
                     i.getArgument(1, Pageable.class),
                     1));
-    var r = _requestFromUri("http://fonzy.com/r4/Questionnaire" + query);
+    var r = requestFromUri("http://fonzy.com/r4/Questionnaire" + query);
     var actual = controller.search(r);
     assertThat(actual.entry()).hasSize(1);
   }
