+++ patient-generated-data/src/test/java/gov/va/api/health/patientgenerateddata/questionnaireresponse/QuestionnaireResponseControllerTest.java
@@ -1,5 +1,6 @@
 package gov.va.api.health.patientgenerateddata.questionnaireresponse;
 
+import static gov.va.api.health.patientgenerateddata.RequestUtils.requestFromUri;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -29,9 +30,7 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.ResponseEntity;
-import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.validation.DataBinder;
-import org.springframework.web.util.UriComponentsBuilder;
 
 public class QuestionnaireResponseControllerTest {
   private static final ObjectMapper MAPPER = JacksonMapperConfig.createMapper();
@@ -43,19 +42,6 @@ public class QuestionnaireResponseControllerTest {
           .baseUrl("http://foo.com")
           .r4BasePath("r4")
           .build();
-
-  private static MockHttpServletRequest requestFromUri(String uri) {
-    var u = UriComponentsBuilder.fromUriString(uri).build();
-    MockHttpServletRequest request = new MockHttpServletRequest();
-    request.setRequestURI(u.getPath());
-    request.setRemoteHost(u.getHost());
-    request.setProtocol(u.getScheme());
-    request.setServerPort(u.getPort());
-    u.getQueryParams()
-        .entrySet()
-        .forEach(e -> request.addParameter(e.getKey(), e.getValue().toArray(new String[0])));
-    return request;
-  }
 
   private QuestionnaireResponseController controller() {
     QuestionnaireResponseRepository repo = mock(QuestionnaireResponseRepository.class);
